requirements.txt
pyodbc
fastapi
uvicorn[standard]
pydantic
azure-identity
python-dotenv
azure-storage-blob 
azure-identity
pandas
pyarrow
